+++ src/pages/RollingPaperPage.module.scss
@@ -1,5 +1,3 @@
-@import "styles/button";
-
 .page-main {
   @media screen and (max-width: 768px) {
     min-width: 360px;
@@ -55,38 +53,5 @@
 
   @media screen and (max-width: 768px) {
     grid-template-columns: 1fr;
-  }
-}
-
-.select-all-container {
-  display: flex;
-  gap: 10px;
-  align-items: center;
-
-  input {
-    width: 30px;
-    height: 30px;
-    cursor: pointer;
-  }
-}
-
-.button-list-container {
-  margin-bottom: 15px;
-  display: flex;
-  justify-content: space-between;
-
-  .delete-paper-button {
-    @include btn-red;
-    @include btn-small;
-    color: var(--white);
-  }
-
-  .go-back-button {
-    background: none;
-  }
-
-  .checkbox-button-container {
-    display: flex;
-    gap: 20px;
   }
 }
